% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepkoski-package.R
\docType{package}
\name{sepkoski-package}
\alias{sepkoski-package}
\title{sepkoski: Sepkoski's Fossil Marine Animal Genera Compendium}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Stratigraphic ranges of fossil marine animal genera from Sepkoski's (2002) published compendium. No changes have been made to any taxonomic names. However, first and last appearance intervals have been updated to be consistent with stages of the International Geological Timescale. Functionality for generating a plot of Sepkoski's evolutionary fauna is also included. For specific details on the compendium see: Sepkoski, J. J. (2002). A compendium of fossil marine animal genera. Bulletins of American Paleontology, 363, pp. 1–560 (ISBN 0-87710-450-6). Access: \url{https://www.biodiversitylibrary.org/item/40634#page/5/mode/1up}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/palaeoverse/sepkoski}
  \item \url{https://palaeoverse.github.io/sepkoski/}
  \item Report bugs at \url{https://github.com/palaeoverse/sepkoski/issues}
}

}
\author{
\strong{Maintainer}: Lewis A. Jones \email{LewisA.Jones@outlook.com} (\href{https://orcid.org/0000-0003-3902-8986}{ORCID})

}
\keyword{internal}
