% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepkoski_curve_base.R
\name{sepkoski_curve_base}
\alias{sepkoski_curve_base}
\title{Plot Sepkoski's evolutionary fauna using base R}
\usage{
sepkoski_curve_base(plot_args = NULL, axis_args = NULL, legend_args = NULL)
}
\arguments{
\item{plot_args}{\code{list}. A named list of optional arguments that are
passed directly to \code{\link[graphics:plot.default]{graphics::plot()}}. If \code{NULL} (default), a default
list of arguments are used.}

\item{axis_args}{\code{list}. A named list of optional arguments that are
passed directly to \code{\link[palaeoverse:axis_geo]{palaeoverse::axis_geo()}}. If \code{NULL} (default), a
default list of arguments are used. If \code{FALSE}, no axis is added.}

\item{legend_args}{\code{list}. A named list of optional arguments that are
passed directly to \code{\link[graphics:legend]{graphics::legend()}}. If \code{NULL} (default), a default
list of arguments are used. If \code{FALSE}, no legend is added.}
}
\value{
No return value. Function is used to plot Sepkoski's curve with
user-defined arguments.
}
\description{
This function plots Sepkoski's evolutionary fauna (Sepkoski, 1981), using the
Sepkoski (2002) fossil marine animal genera compendium (i.e. the included
\link[sepkoski]{sepkoski} dataset). No changes have been made to any
taxonomic names in this dataset. However, first and last appearance
intervals have been updated to stages from the
\href{https://stratigraphy.org/ICSchart/ChronostratChart2023-09.pdf}{International Geological Time Scale 2023}. As
such, minor differences may be observed to previously published plots.
}
\details{
Taxa are assigned to evolutionary fauna (EF) categories as follows:
\itemize{
\item Cambrian EF: Trilobita, Polychaeta, Tergomya ("Monoplacophora"),
Inarticulata, and Hyolithomorpha.
\item Paleozoic EF: Anthozoa, Articulata, Asteroidea, Cephalopoda, Crinoidea,
Ostracoda, Ophiuroidea, Somasteroidea, and Stenolaemata.
\item Modern EF: Bivalvia, Chondrichthyes, Demospongia, Echinoidea, Gastropoda,
Gymnolaemata, Malacostraca, and Osteichthyes.
}
}
\section{References}{

Sepkoski, J. J. (1981).  A factor analytic description of the Phanerozoic
marine fossil record. \emph{Paleobiology}, 7(1), pp. 36--53.

Sepkoski, J. J. (2002). A compendium of fossil marine animal genera.
\emph{Bulletins of American Paleontology}, 363, pp. 1--560.
}

\examples{
# Plot curve with default arguments
sepkoski_curve_base()

# Plot curve with user-defined arguments
sepkoski_curve_base(plot_args = list(main = "Sepkoski's Curve"),
                    axis_args = list(intervals = list("stages", "periods"),
                                    lab = FALSE),
                    legend_args = list(bty = "o"))
}
