% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepkoski-package.R
\docType{package}
\name{sepkoski-package}
\alias{sepkoski-package}
\title{sepkoski: Sepkoski's Fossil Marine Animal Genera Compendium}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Stratigraphic ranges of fossil marine animal genera from Sepkoski's (2002) published compendium. This version of the dataset is a port of Shanan Peter's online database (\url{http://strata.geology.wisc.edu/jack/}). No changes have been made to any taxonomic names. However, first and last appearance intervals have been updated to be consistent with stages of the International Geological Timescale 2022. Functionality for generating a plot of Sepkoski's evolutionary fauna is also included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LewisAJones/sepkoski}
  \item Report bugs at \url{https://github.com/LewisAJones/sepkoski/issues}
}

}
\author{
\strong{Maintainer}: Lewis A. Jones \email{LewisA.Jones@outlook.com} (\href{https://orcid.org/0000-0003-3902-8986}{ORCID})

}
\keyword{internal}
